Improve logic to derive branch name from Gogs webhook ref
Branches often times contain the / character, so instead of interpreting
the text after the last / we remove `refs/heads/` from the ref.

Also, Jenkins URL encodes job names internally. So a multi branch job
name for a branch named `feature/test-branch` is internally named
`feature%2Ftest-branch`. To account for this the derived branchName is
URL encoded before passing it to findJob.